Wrongly interpretation of vsnprintf return code
Formatting a message after we received a -1 can cause segfaults. This fix properly implements the needed buffer size check and the additional allocation afterwards. The fix could be more efficient if we could use a REALLOC. The code will also stop gracefully whenever vsnprintf cannot parse the message.

It also fixed a potential memory leak because the buffer was overwritten by a new malloc instead of resized by a realloc.